# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/overview": {
      "get": {
        "operationId": "AdminController_overview",
        "parameters": [
          {
            "name": "X-Admin-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Current releases and rollout summary"
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "summary": "Release operations dashboard summary",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/releases": {
      "get": {
        "operationId": "AdminController_list",
        "parameters": [
          {
            "name": "X-Admin-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminController_create",
        "parameters": [
          {
            "name": "X-Admin-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Release metadata and optional verified artifact",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "applicationId",
                  "platform",
                  "version",
                  "buildNumber",
                  "runtimeVersion",
                  "channel"
                ],
                "properties": {
                  "applicationId": {
                    "type": "string"
                  },
                  "platform": {
                    "enum": [
                      "android",
                      "ios"
                    ]
                  },
                  "version": {
                    "type": "string"
                  },
                  "buildNumber": {
                    "type": "integer"
                  },
                  "runtimeVersion": {
                    "type": "string"
                  },
                  "channel": {
                    "enum": [
                      "store",
                      "direct",
                      "mdm",
                      "ota"
                    ]
                  },
                  "releaseNotes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "artifact": {
                    "type": "object",
                    "nullable": true
                  },
                  "rollout": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/releases/{id}": {
      "get": {
        "operationId": "AdminController_detail",
        "parameters": [
          {
            "name": "X-Admin-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/releases/{id}/{action}": {
      "post": {
        "operationId": "AdminController_action",
        "parameters": [
          {
            "name": "X-Admin-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "action",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "High-risk action confirmation",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason",
                  "confirm"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3
                  },
                  "confirm": {
                    "type": "boolean",
                    "enum": [
                      true
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/audit-events": {
      "get": {
        "operationId": "AdminController_audits",
        "parameters": [
          {
            "name": "X-Admin-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Admin-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "rn_admin_session": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "RN Foundation API",
    "description": "Mobile bootstrap, localization, theme and update policy",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rn_admin_session"
      }
    },
...
  }
}